backend/requirements.txt
# =========================
# CORE BACKEND
# =========================
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0

# =========================
# DATABASE
# =========================
sqlalchemy>=2.0
sqlmodel>=0.0.16

# =========================
# HTTP / UTILIDADES
# =========================
httpx>=0.27
pydantic>=2.6

# =========================
# PLAYWRIGHT (PRERENDER SEO)
# =========================
playwright>=1.44

# =========================
# OPCIONAL (pero recomendado)
# =========================
jinja2>=3.1
